refactor: more net refactor

+++ src/main/kotlin/kyta/composter/helpers.kt
@@ -10,6 +10,6 @@ fun Double.asAbsoluteInt(): Int {
     return (this * 32.0).toInt()
 }
 
-fun Float.asRotation(): Float {
-    return this * 256F / 360F
+fun Float.asRotation(): Int {
+    return (this * 256F / 360F).toInt()
 }
